+++ src/ShellSdk.ts
@@ -2,6 +2,14 @@ import { EventType, ErrorType, SHELL_EVENTS } from './ShellEvents';
 import { SHELL_VERSION_INFO } from './ShellVersionInfo';
 import { Debugger } from './Debugger';
 import { Outlet } from './models/outlets/outlet.model';
+import { PayloadValidator } from './validation/interfaces/payload-validator';
+import {
+  getEventValidationConfiguration,
+  EventValidationConfiguration,
+} from './validation/schemas/validation-configuration';
+import { PayloadValidationError } from './validation/payload-validation-error';
+
+export type ValidationMode = 'host' | 'client';
 
 // tslint:disable
 function uuidv4() {
@@ -22,6 +30,9 @@ export class ShellSdk {
   private static _instance: ShellSdk;
   private isRoot: boolean; // Is root if on `init`, target value is null.
   private isInsideModal: boolean;
+  private validator: null | PayloadValidator = null;
+  private validationMode: ValidationMode = 'client';
+  private eventValidationConfiguration: EventValidationConfiguration;
 
   private postMessageHandler:
     | (<T>(type: EventType, value: T, to?: string[]) => void)
@@ -52,6 +63,7 @@ export class ShellSdk {
     this.debugger = new Debugger(winRef, debugId);
     this.isRoot = target == null;
     this.isInsideModal = false;
+    this.eventValidationConfiguration = getEventValidationConfiguration();
   }
 
   public static init(
@@ -162,6 +174,14 @@ export class ShellSdk {
     );
   }
 
+  public setValidator(
+    validator: PayloadValidator,
+    validationMode: ValidationMode = 'client'
+  ) {
+    this.validator = validator;
+    this.validationMode = validationMode;
+  }
+
   // Called by outlet component to assign an generated uuid to an iframe. This is key
   // to allow one to one communication between a pluging and shell-host
   public registerOutlet(frame: HTMLIFrameElement, _name: string | undefined) {
@@ -227,6 +247,28 @@ export class ShellSdk {
     if (!this.postMessageHandler) {
       throw new Error("ShellSdk wasn't initialized, message handler not set.");
     }
+
+    if (!!this.validator && !!this.eventValidationConfiguration[type]) {
+      const validationConfig =
+        this.validationMode === 'client'
+          ? this.eventValidationConfiguration[type].request
+          : this.eventValidationConfiguration[type].response;
+
+      if (!!validationConfig) {
+        if (!validationConfig.validationFunction) {
+          validationConfig.validationFunction =
+            this.validator.getValidationFunction(validationConfig.schema);
+        }
+        const validationResult = validationConfig.validationFunction(value);
+        if (!validationResult.isValid) {
+          throw new PayloadValidationError(
+            'Payload validation failed',
+            validationResult.error
+          );
+        }
+      }
+    }
+
     // Only root can send a message to a specific node
     this.postMessageHandler(type, value, this.isRoot ? to : undefined);
   }