+++ src/validation/payload-validation-error.ts
@@ -1,0 +1,9 @@
+export class PayloadValidationError extends Error {
+  public detail: any;
+
+  constructor(message: string, detail: any) {
+    super(message);
+    this.name = 'PayloadValidationError';
+    this.detail = detail;
+  }
+}