+++ src/validation/schemas/authentication/auth-request.v1.schema.ts
@@ -1,0 +1,10 @@
+export const authRequest_v1_schema = {
+  type: 'object',
+  properties: {
+    response_type: {
+      type: 'string',
+      enum: ['token', 'code'],
+    },
+  },
+  required: ['response_type'],
+};