+++ src/validation/schemas/cloud-storage/get-item-request.v2.schema.ts
@@ -1,0 +1,3 @@
+export const getItemRequest_v2_schema = {
+  type: 'string',
+};