+++ src/validation/schemas/cloud-storage/get-item-response.v2.schema.ts
@@ -1,0 +1,10 @@
+export const getItemResponse_v2_schema = {
+  type: 'object',
+  properties: {
+    key: {
+      type: 'string',
+    },
+    value: {},
+  },
+  required: ['key', 'value'],
+};