+++ src/validation/schemas/outlets/outlets-request-context-response.v1.schema.ts
@@ -1,0 +1,16 @@
+export const outletsRequestContextResponse_v1_schema = {
+  type: 'object',
+  properties: {
+    target: {
+      type: 'string',
+    },
+    isRootNodeHttps: {
+      type: 'boolean',
+    },
+    isConfigurationMode: {
+      type: 'boolean',
+    },
+    plugin: {},
+  },
+  required: ['target', 'isRootNodeHttps'],
+};