+++ src/validation/schemas/permissions/get-permissions-request.v2.schema.ts
@@ -1,0 +1,15 @@
+export const getPermissionsRequest_v2_schema = {
+  type: 'object',
+  properties: {
+    objectName: {
+      type: 'string',
+    },
+    owners: {
+      type: 'array',
+      items: {
+        type: 'string',
+      },
+    },
+  },
+  required: ['objectName'],
+};