+++ src/validation/schemas/view-state/set-view-state-request.v1.schema.ts
@@ -1,0 +1,10 @@
+export const setViewStateRequest_v1_schema = {
+  type: 'object',
+  properties: {
+    key: {
+      type: 'string',
+    },
+    value: {},
+  },
+  required: ['key', 'value'],
+};