+++ CHANGELOG.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [1.15.0] - 2021-11-17
 
 ### Added
 
 - Enhance fsm-shell to support dynamic outlets
+
+## [1.14.0] - 2021-10-21
+
+### Added
+
+- Json schemas for event payload validation
+- Logic for event payload validation against json schemas
 
 ## [1.13.2] - 2021-09-24